Revert "Add http GET endpoint for accessing project snapshot (#5546)"
This reverts commit 6cfb0e427e1800e74a83c9fe9e20b8ff5d657fc2.
