+++ tests/inspection/test_project_snapshot.py
@@ -186,24 +186,14 @@ class TestBuildProjectSnapshot:
     def test_config_loader_created_with_explicit_env(self):
         _build_project_snapshot(self.project_path, env="staging")
         self.mock_make_config_loader.assert_called_once_with(
-            self.project_path, env="staging", conf_source=None
+            self.project_path, env="staging"
         )
 
     def test_config_loader_created_with_default_env(self):
         _build_project_snapshot(self.project_path)
         self.mock_make_config_loader.assert_called_once_with(
-            self.project_path, env=None, conf_source=None
+            self.project_path, env=None
         )
-
-    def test_config_loader_created_with_explicit_conf_source(self):
-        _build_project_snapshot(self.project_path, conf_source="conf/custom")
-        self.mock_make_config_loader.assert_called_once_with(
-            self.project_path, env=None, conf_source="conf/custom"
-        )
-
-    def test_metadata_parameter_skips_bootstrap(self, project_metadata):
-        _build_project_snapshot(self.project_path, metadata=project_metadata)
-        self.mock_bootstrap.assert_not_called()
 
     def test_catalog_loaded_from_config_loader_and_passed_downstream(self, mocker):
         """Catalog config is loaded from the config loader and passed to both
@@ -282,34 +272,3 @@ class TestBuildProjectSnapshot:
     )
     def test_valid_env_does_not_raise(self, env):
         _build_project_snapshot(self.project_path, env=env)
-
-    def test_raises_when_neither_project_path_nor_metadata_provided(self):
-        with pytest.raises(
-            ValueError, match="Either project_path or metadata must be provided"
-        ):
-            _build_project_snapshot()
-
-    def test_warns_when_project_path_and_metadata_project_path_differ(
-        self, project_metadata, tmp_path
-    ):
-        different_path = tmp_path / "other_project"
-        with pytest.warns(UserWarning, match="project_path will be ignored"):
-            _build_project_snapshot(different_path, metadata=project_metadata)
-
-    def test_no_warning_when_project_path_matches_metadata(self, project_metadata):
-        import warnings
-
-        with warnings.catch_warnings():
-            warnings.simplefilter("error")
-            _build_project_snapshot(
-                project_metadata.project_path, metadata=project_metadata
-            )
-
-    def test_metadata_project_path_used_for_config_loader_when_only_metadata_provided(
-        self, project_metadata
-    ):
-        _build_project_snapshot(metadata=project_metadata)
-        self.mock_bootstrap.assert_not_called()
-        self.mock_make_config_loader.assert_called_once_with(
-            project_metadata.project_path, env=None, conf_source=None
-        )
